backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
google-generativeai==0.8.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.1
python-docx==1.0.1
pypdf==3.16.4
beautifulsoup4==4.12.2
python-dotenv==1.0.0
numpy<2.0.0
pytest
# For Python < 3.10 compatibility
importlib-metadata>=8.7.1
zipp>=3.23.0
